Rename algorithms.py to tidysic.py
