+++ tidysic/__init__.py
@@ -1,4 +1,4 @@
-from .algorithms import (
+from .tidysic import (
     create_structure,
     move_files,
     clean_up,
